fix(db): un archivo por entidad en schema/utils, no taxonomia.ts
server/db/schema/ y server/utils/ agrupaban categorias y comunas en
un solo archivo — CLAUDE.md (A-006) pide un archivo por entidad, sin
excepción, aunque las dos tablas nazcan de la misma misión. Se
dividen en categorias.ts/comunas.ts en ambas carpetas. Sin cambio de
schema real (drizzle-kit generate confirma "no changes"), solo
reorganización de archivos — ingenieria.md actualizado para reflejar
los nombres correctos.
